combine start and end date pages
