+++ TeachingRecordSystem/src/TeachingRecordSystem.SupportUi/Pages/RoutesToProfessionalStatus/AddRoutePage.cs
@@ -4,8 +4,7 @@ public enum AddRoutePage
 {
     Route = 0,
     Status,
-    StartDate,
-    EndDate,
+    StartAndEndDate,
     AwardDate,
     InductionExemption,
     TrainingProvider,
