+++ TeachingRecordSystem/src/TeachingRecordSystem.SupportUi/Pages/RoutesToProfessionalStatus/EditRoute/EditRouteStatusState.cs
@@ -4,7 +4,6 @@ public class EditRouteStatusState
 {
     public RouteToProfessionalStatusStatus Status { get; set; }
     public DateOnly? HoldsFrom { get; set; }
-    public DateOnly? TrainingEndDate { get; set; }
     public bool? InductionExemption { get; set; }
     public bool RouteImplicitExemption { get; init; }
 }
