+++ TeachingRecordSystem/src/TeachingRecordSystem.SupportUi/Pages/RoutesToProfessionalStatus/EditRoute/HoldsFrom.cshtml.cs
@@ -54,7 +54,6 @@ public class HoldsFromModel(TrsLinkGenerator linkGenerator) : PageModel
                 await JourneyInstance!.UpdateStateAsync(s => // update the main journey state with the data
                 {
                     s.Status = s.EditStatusState!.Status;
-                    s.TrainingEndDate = s.EditStatusState.TrainingEndDate.HasValue ? s.EditStatusState.TrainingEndDate.Value : s.TrainingEndDate;
                     s.HoldsFrom = HoldsFrom;
                     s.IsExemptFromInduction = s.EditStatusState.InductionExemption;
                     s.EditStatusState = null;
@@ -126,6 +125,6 @@ public class HoldsFromModel(TrsLinkGenerator linkGenerator) : PageModel
 
     private string PreviousCompletingRoutePage =>
         QuestionDriverHelper.FieldRequired(RouteToProfessionalStatus!.TrainingEndDateRequired, JourneyInstance!.State.EditStatusState!.Status.GetEndDateRequirement()) != FieldRequirement.NotApplicable ?
-            linkGenerator.RouteEditEndDate(QualificationId, JourneyInstance!.InstanceId) :
+            linkGenerator.RouteEditStartAndEndDate(QualificationId, JourneyInstance!.InstanceId) :
             linkGenerator.RouteEditStatus(QualificationId, JourneyInstance!.InstanceId);
 }
