+++ TeachingRecordSystem/src/TeachingRecordSystem.SupportUi/Pages/RoutesToProfessionalStatus/EditRoute/StartAndEndDates.cshtml.cs
@@ -6,7 +6,7 @@ using Microsoft.AspNetCore.Mvc.RazorPages;
 namespace TeachingRecordSystem.SupportUi.Pages.RoutesToProfessionalStatus.EditRoute;
 
 [Journey(JourneyNames.EditRouteToProfessionalStatus), RequireJourneyInstance]
-public class StartDateModel(
+public class StartAndEndDateModel(
     TrsLinkGenerator linkGenerator) : PageModel
 {
     public JourneyInstance<EditRouteState>? JourneyInstance { get; set; }
@@ -24,34 +24,53 @@ public class StartDateModel(
     [BindProperty]
     [DateInput(ErrorMessagePrefix = "Start date")]
     [Required(ErrorMessage = "Enter a start date")]
-    [Display(Name = "Enter the route start date")]
+    [Display(Name = "Route start date")]
     public DateOnly? TrainingStartDate { get; set; }
+
+    [BindProperty]
+    [DateInput(ErrorMessagePrefix = "End date")]
+    [Required(ErrorMessage = "Enter an end date")]
+    [Display(Name = "Route end date")]
+    public DateOnly? TrainingEndDate { get; set; }
 
     public void OnGet()
     {
         TrainingStartDate = JourneyInstance!.State.TrainingStartDate;
+        TrainingEndDate = JourneyInstance!.State.TrainingEndDate;
     }
 
     public async Task<IActionResult> OnPostAsync()
     {
+        if (TrainingStartDate >= TrainingEndDate)
+        {
+            ModelState.AddModelError(nameof(TrainingEndDate), "End date must be after start date");
+        }
         if (!ModelState.IsValid)
         {
             return this.PageWithErrors();
         }
-        await JourneyInstance!.UpdateStateAsync(state => state.TrainingStartDate = TrainingStartDate);
-        if (TrainingStartDate.HasValue && JourneyInstance!.State.TrainingEndDate is DateOnly endDate && TrainingStartDate >= endDate)
+        await JourneyInstance!.UpdateStateAsync(s =>
         {
-            return Redirect(linkGenerator.RouteEditEndDate(QualificationId, JourneyInstance.InstanceId));
-        }
-        return Redirect(FromCheckAnswers ?
-            linkGenerator.RouteCheckYourAnswers(QualificationId, JourneyInstance.InstanceId) :
-            linkGenerator.RouteEditDetail(QualificationId, JourneyInstance.InstanceId));
+            s.TrainingStartDate = TrainingStartDate;
+            s.TrainingEndDate = TrainingEndDate;
+        });
+
+        return Redirect(JourneyInstance!.State.IsCompletingRoute ?
+            NextCompletingRoutePage() :
+            FromCheckAnswers ?
+                linkGenerator.RouteCheckYourAnswers(QualificationId, JourneyInstance.InstanceId) :
+                linkGenerator.RouteEditDetail(QualificationId, JourneyInstance.InstanceId));
     }
 
     public async Task<IActionResult> OnPostCancelAsync()
     {
         await JourneyInstance!.DeleteAsync();
         return Redirect(linkGenerator.PersonQualifications(PersonId));
+    }
+
+    private string NextCompletingRoutePage()
+    {
+        return linkGenerator.RouteEditHoldsFrom(QualificationId, JourneyInstance!.InstanceId);
     }
 
     public override void OnPageHandlerExecuting(PageHandlerExecutingContext context)
@@ -61,3 +80,4 @@ public class StartDateModel(
         PersonId = personInfo.PersonId;
     }
 }
+
