+++ TeachingRecordSystem/tests/TeachingRecordSystem.SupportUi.Tests/PageTests/RoutesToProfessionalStatus/AddRoute/StartAndEndDateTests.cs
@@ -1,10 +1,9 @@
-using AngleSharp.Dom;
 using AngleSharp.Html.Dom;
 using TeachingRecordSystem.SupportUi.Pages.RoutesToProfessionalStatus.AddRoute;
 
 namespace TeachingRecordSystem.SupportUi.Tests.PageTests.RoutesToProfessionalStatus.AddRoute;
 
-public class EndDateTests(HostFixture hostFixture) : TestBase(hostFixture)
+public class StartAndEndDateTests(HostFixture hostFixture) : TestBase(hostFixture)
 {
     [Fact]
     public async Task Get_ShowsPreviouslyStoredEntry()
@@ -32,21 +31,24 @@ public class EndDateTests(HostFixture hostFixture) : TestBase(hostFixture)
             addRouteState
             );
 
-        var request = new HttpRequestMessage(HttpMethod.Get, $"/route/add/end-date?personId={person.PersonId}&{journeyInstance.GetUniqueIdQueryParameter()}");
+        var request = new HttpRequestMessage(HttpMethod.Get, $"/route/add/start-and-end-date?personId={person.PersonId}&{journeyInstance.GetUniqueIdQueryParameter()}");
 
         // Act
         var response = await HttpClient.SendAsync(request);
 
         // Assert
         var doc = await AssertEx.HtmlResponseAsync(response);
-        var displayedDate = doc.QuerySelectorAll<IHtmlInputElement>("[type=text]");
-        Assert.Equal(endDate.Day.ToString(), displayedDate.ElementAt(0).Value);
-        Assert.Equal(endDate.Month.ToString(), displayedDate.ElementAt(1).Value);
-        Assert.Equal(endDate.Year.ToString(), displayedDate.ElementAt(2).Value);
+
+        Assert.Equal(startDate.Day.ToString(), doc.QuerySelector("#TrainingStartDate\\.Day")?.GetAttribute("value"));
+        Assert.Equal(startDate.Month.ToString(), doc.QuerySelector("#TrainingStartDate\\.Month")?.GetAttribute("value"));
+        Assert.Equal(startDate.Year.ToString(), doc.QuerySelector("#TrainingStartDate\\.Year")?.GetAttribute("value"));
+        Assert.Equal(endDate.Day.ToString(), doc.QuerySelector("#TrainingEndDate\\.Day")?.GetAttribute("value"));
+        Assert.Equal(endDate.Month.ToString(), doc.QuerySelector("#TrainingEndDate\\.Month")?.GetAttribute("value"));
+        Assert.Equal(endDate.Year.ToString(), doc.QuerySelector("#TrainingEndDate\\.Year")?.GetAttribute("value"));
     }
 
     [Fact]
-    public async Task Post_WhenEndDateIsEntered_SavesDateAndRedirectsToNextPage()
+    public async Task Post_WhenStartAndEndDateIsEntered_SavesDateAndRedirectsToNextPage()
     {
         // Arrange
         var startDate = new DateOnly(2024, 01, 01);
@@ -62,7 +64,6 @@ public class EndDateTests(HostFixture hostFixture) : TestBase(hostFixture)
         var addRouteState = new AddRouteStateBuilder()
             .WithRouteToProfessionalStatusId(route.RouteToProfessionalStatusTypeId)
             .WithStatus(status)
-            .WithTrainingStartDate(startDate)
             .Build();
 
         var journeyInstance = await CreateJourneyInstanceAsync(
@@ -70,10 +71,13 @@ public class EndDateTests(HostFixture hostFixture) : TestBase(hostFixture)
             addRouteState
             );
 
-        var request = new HttpRequestMessage(HttpMethod.Post, $"/route/add/end-date?personId={person.PersonId}&{journeyInstance.GetUniqueIdQueryParameter()}")
+        var request = new HttpRequestMessage(HttpMethod.Post, $"/route/add/start-and-end-date?personId={person.PersonId}&{journeyInstance.GetUniqueIdQueryParameter()}")
         {
             Content = new FormUrlEncodedContentBuilder()
             {
+                { "TrainingStartDate.Day", $"{startDate:%d}" },
+                { "TrainingStartDate.Month", $"{startDate:%M}" },
+                { "TrainingStartDate.Year", $"{startDate:yyyy}" },
                 { "TrainingEndDate.Day", $"{endDate:%d}" },
                 { "TrainingEndDate.Month", $"{endDate:%M}" },
                 { "TrainingEndDate.Year", $"{endDate:yyyy}" },
@@ -117,13 +121,16 @@ public class EndDateTests(HostFixture hostFixture) : TestBase(hostFixture)
             addRouteState
             );
 
-        var request = new HttpRequestMessage(HttpMethod.Post, $"/route/add/end-date?personId={person.PersonId}&FromCheckAnswers=True&{journeyInstance.GetUniqueIdQueryParameter()}")
+        var request = new HttpRequestMessage(HttpMethod.Post, $"/route/add/start-and-end-date?personId={person.PersonId}&FromCheckAnswers=True&{journeyInstance.GetUniqueIdQueryParameter()}")
         {
             Content = new FormUrlEncodedContentBuilder()
             {
+                { "TrainingStartDate.Day", $"{startDate:%d}" },
+                { "TrainingStartDate.Month", $"{startDate:%M}" },
+                { "TrainingStartDate.Year", $"{startDate:yyyy}" },
                 { "TrainingEndDate.Day", $"{newEndDate:%d}" },
                 { "TrainingEndDate.Month", $"{newEndDate:%M}" },
-                { "TrainingEndDate.Year", $"{newEndDate:yyyy}" },
+                { "TrainingEndDate.Year", $"{newEndDate:yyyy}" }
             }
         };
 
@@ -135,6 +142,47 @@ public class EndDateTests(HostFixture hostFixture) : TestBase(hostFixture)
         Assert.Equal($"/route/add/check-answers?personId={person.PersonId}&{journeyInstance.GetUniqueIdQueryParameter()}", response.Headers.Location?.OriginalString);
         journeyInstance = await ReloadJourneyInstance(journeyInstance);
         Assert.Equal(newEndDate, journeyInstance.State.TrainingEndDate);
+    }
+
+    [Fact]
+    public async Task Post_WhenNoStartDateIsEntered_ReturnsError()
+    {
+        // Arrange
+        var startDate = new DateOnly(2024, 01, 01);
+        var endDate = startDate.AddYears(1);
+        var route = (await ReferenceDataCache.GetRouteToProfessionalStatusTypesAsync())
+            .Where(r => r.TrainingStartDateRequired == FieldRequirement.Mandatory && r.TrainingEndDateRequired == FieldRequirement.Mandatory)
+            .RandomOne();
+        var status = ProfessionalStatusStatusRegistry.All
+            .Where(s => s.TrainingStartDateRequired == FieldRequirement.Mandatory && s.TrainingEndDateRequired == FieldRequirement.Mandatory)
+            .RandomOne()
+            .Value;
+        var person = await TestData.CreatePersonAsync();
+
+        var addRouteState = new AddRouteStateBuilder()
+            .WithRouteToProfessionalStatusId(route.RouteToProfessionalStatusTypeId)
+            .WithStatus(status)
+            .Build();
+
+        var journeyInstance = await CreateJourneyInstanceAsync(
+            person.PersonId,
+            addRouteState
+            );
+
+        var request = new HttpRequestMessage(HttpMethod.Post, $"/route/add/start-and-end-date?personId={person.PersonId}&{journeyInstance.GetUniqueIdQueryParameter()}")
+        {
+            Content = new FormUrlEncodedContentBuilder()
+            {
+                { "TrainingEndDate.Day", $"{endDate:%d}" },
+                { "TrainingEndDate.Month", $"{endDate:%M}" },
+                { "TrainingEndDate.Year", $"{endDate:yyyy}" },
+            }
+        };
+        // Act
+        var response = await HttpClient.SendAsync(request);
+
+        // Assert
+        await AssertEx.HtmlResponseHasErrorAsync(response, "TrainingStartDate", "Enter a start date");
     }
 
     [Fact]
@@ -155,7 +203,6 @@ public class EndDateTests(HostFixture hostFixture) : TestBase(hostFixture)
         var addRouteState = new AddRouteStateBuilder()
             .WithRouteToProfessionalStatusId(route.RouteToProfessionalStatusTypeId)
             .WithStatus(status)
-            .WithTrainingStartDate(startDate)
             .Build();
 
         var journeyInstance = await CreateJourneyInstanceAsync(
@@ -163,9 +210,14 @@ public class EndDateTests(HostFixture hostFixture) : TestBase(hostFixture)
             addRouteState
             );
 
-        var request = new HttpRequestMessage(HttpMethod.Post, $"/route/add/end-date?personId={person.PersonId}&{journeyInstance.GetUniqueIdQueryParameter()}")
+        var request = new HttpRequestMessage(HttpMethod.Post, $"/route/add/start-and-end-date?personId={person.PersonId}&{journeyInstance.GetUniqueIdQueryParameter()}")
         {
-            Content = new FormUrlEncodedContent(new Dictionary<string, string>())
+            Content = new FormUrlEncodedContentBuilder()
+            {
+                { "TrainingStartDate.Day", $"{startDate:%d}" },
+                { "TrainingStartDate.Month", $"{startDate:%M}" },
+                { "TrainingStartDate.Year", $"{startDate:yyyy}" }
+            }
         };
 
         // Act
@@ -201,13 +253,16 @@ public class EndDateTests(HostFixture hostFixture) : TestBase(hostFixture)
             addRouteState
             );
 
-        var request = new HttpRequestMessage(HttpMethod.Post, $"/route/add/end-date?personId={person.PersonId}&{journeyInstance.GetUniqueIdQueryParameter()}")
+        var request = new HttpRequestMessage(HttpMethod.Post, $"/route/add/start-and-end-date?personId={person.PersonId}&{journeyInstance.GetUniqueIdQueryParameter()}")
         {
             Content = new FormUrlEncodedContentBuilder()
             {
+                { "TrainingStartDate.Day", $"{startDate:%d}" },
+                { "TrainingStartDate.Month", $"{startDate:%M}" },
+                { "TrainingStartDate.Year", $"{startDate:yyyy}" },
                 { "TrainingEndDate.Day", $"{endDate:%d}" },
                 { "TrainingEndDate.Month", $"{endDate:%M}" },
-                { "TrainingEndDate.Year", $"{endDate:yyyy}" },
+                { "TrainingEndDate.Year", $"{endDate:yyyy}" }
             }
         };
 
@@ -240,7 +295,7 @@ public class EndDateTests(HostFixture hostFixture) : TestBase(hostFixture)
             addRouteState
             );
 
-        var request = new HttpRequestMessage(HttpMethod.Get, $"/route/add/end-date?personId={person.PersonId}&{journeyInstance.GetUniqueIdQueryParameter()}");
+        var request = new HttpRequestMessage(HttpMethod.Get, $"/route/add/start-and-end-date?personId={person.PersonId}&{journeyInstance.GetUniqueIdQueryParameter()}");
 
         // Act
         var response = await HttpClient.SendAsync(request);
