--- TeachingRecordSystem/tests/TeachingRecordSystem.SupportUi.Tests/PageTests/RoutesToProfessionalStatus/EditRoute/EndDateTests.cs
@@ -1,252 +1,0 @@
-using AngleSharp.Html.Dom;
-using TeachingRecordSystem.SupportUi.Pages.RoutesToProfessionalStatus.EditRoute;
-
-namespace TeachingRecordSystem.SupportUi.Tests.PageTests.RoutesToProfessionalStatus.EditRoute;
-
-public class EndDateTests(HostFixture hostFixture) : TestBase(hostFixture)
-{
-    [Theory]
-    [InlineData(0)]
-    [InlineData(-1)]
-    public async Task Post_WhenEndDateIsEqualOrBeforeStartDate_RendersError(int daysAfter)
-    {
-        // Arrange
-        var startDate = new DateOnly(2024, 01, 01);
-        var endDate = startDate.AddDays(daysAfter);
-        var route = (await ReferenceDataCache.GetRouteToProfessionalStatusTypesAsync())
-            .Where(r => r.TrainingEndDateRequired == FieldRequirement.Mandatory)
-            .RandomOne();
-        var status = ProfessionalStatusStatusRegistry.All
-            .Where(s => s.TrainingEndDateRequired == FieldRequirement.Mandatory && s.HoldsFromRequired == FieldRequirement.NotApplicable)
-            .RandomOne()
-            .Value;
-        var person = await TestData.CreatePersonAsync(p => p
-            .WithRouteToProfessionalStatus(r => r
-                .WithRouteType(route.RouteToProfessionalStatusTypeId)
-                .WithStatus(status)));
-        var qualificationid = person.ProfessionalStatuses.First().QualificationId;
-        var editRouteState = new EditRouteStateBuilder()
-            .WithRouteToProfessionalStatusId(route.RouteToProfessionalStatusTypeId)
-            .WithStatus(status)
-            .WithTrainingStartDate(startDate)
-            .WithValidChangeReasonOption()
-            .WithDefaultChangeReasonNoUploadFileDetail()
-            .Build();
-
-        var journeyInstance = await CreateJourneyInstanceAsync(
-            qualificationid,
-            editRouteState
-            );
-
-        var request = new HttpRequestMessage(HttpMethod.Post, $"/route/{qualificationid}/edit/end-date?{journeyInstance.GetUniqueIdQueryParameter()}")
-        {
-            Content = new FormUrlEncodedContentBuilder()
-            {
-                { "TrainingEndDate.Day", $"{endDate:%d}" },
-                { "TrainingEndDate.Month", $"{endDate:%M}" },
-                { "TrainingEndDate.Year", $"{endDate:yyyy}" },
-            }
-        };
-
-        // Act
-        var response = await HttpClient.SendAsync(request);
-
-        // Assert
-        await AssertEx.HtmlResponseHasErrorAsync(response, "TrainingEndDate", "End date must be after start date");
-    }
-
-    [Fact]
-    public async Task Post_NotStatusHoldsJourney_TrainingEndDateIsEntered_SavesDateAndRedirectsToDetail()
-    {
-        // Arrange
-        var startDate = new DateOnly(2024, 01, 01);
-        var endDate = new DateOnly(2025, 01, 01);
-        var route = (await ReferenceDataCache.GetRouteToProfessionalStatusTypesAsync())
-            .Where(r => r.TrainingEndDateRequired == FieldRequirement.Mandatory)
-            .RandomOne();
-        var status = ProfessionalStatusStatusRegistry.All
-            .Where(s => s.TrainingEndDateRequired == FieldRequirement.Mandatory
-                && s.Value != RouteToProfessionalStatusStatus.Holds)
-            .RandomOne()
-            .Value;
-        var person = await TestData.CreatePersonAsync(p => p
-            .WithRouteToProfessionalStatus(r => r
-                .WithRouteType(route.RouteToProfessionalStatusTypeId)
-                .WithStatus(status)));
-        var qualificationid = person.ProfessionalStatuses.First().QualificationId;
-        var editRouteState = new EditRouteStateBuilder()
-            .WithRouteToProfessionalStatusId(route.RouteToProfessionalStatusTypeId)
-            .WithStatus(status)
-            .WithTrainingStartDate(startDate)
-            .Build();
-
-        var journeyInstance = await CreateJourneyInstanceAsync(
-            qualificationid,
-            editRouteState
-            );
-
-        var request = new HttpRequestMessage(HttpMethod.Post, $"/route/{qualificationid}/edit/end-date?{journeyInstance.GetUniqueIdQueryParameter()}")
-        {
-            Content = new FormUrlEncodedContentBuilder()
-            {
-                { "TrainingEndDate.Day", $"{endDate:%d}" },
-                { "TrainingEndDate.Month", $"{endDate:%M}" },
-                { "TrainingEndDate.Year", $"{endDate:yyyy}" },
-            }
-        };
-
-        // Act
-        var response = await HttpClient.SendAsync(request);
-
-        // Assert
-        Assert.Equal(StatusCodes.Status302Found, (int)response.StatusCode);
-        Assert.Equal($"/route/{qualificationid}/edit/detail?{journeyInstance.GetUniqueIdQueryParameter()}", response.Headers.Location?.OriginalString);
-        journeyInstance = await ReloadJourneyInstance(journeyInstance);
-        Assert.Equal(endDate, journeyInstance.State.TrainingEndDate);
-    }
-
-    [Fact]
-    public async Task Post_StatusHoldsJourney_TrainingEndDateIsEntered_SavesDateAndRedirectsToHoldsFromDate()
-    {
-        // Arrange
-        var status = RouteToProfessionalStatusStatus.Holds;
-        var startDate = new DateOnly(2024, 01, 01);
-        var endDate = new DateOnly(2025, 01, 01);
-        var route = (await ReferenceDataCache.GetRouteToProfessionalStatusTypesAsync())
-            .Where(r => r.TrainingEndDateRequired == FieldRequirement.Mandatory)
-            .RandomOne();
-        var person = await TestData.CreatePersonAsync(p => p
-            .WithRouteToProfessionalStatus(r => r
-                .WithRouteType(route.RouteToProfessionalStatusTypeId)
-                .WithStatus(RouteToProfessionalStatusStatus.InTraining)));
-        var qualificationid = person.ProfessionalStatuses.First().QualificationId;
-        var editRouteState = new EditRouteStateBuilder()
-            .WithRouteToProfessionalStatusId(route.RouteToProfessionalStatusTypeId)
-            .WithStatus(status)
-            .WithCurrentStatus(person.ProfessionalStatuses.First().Status)
-            .WithTrainingStartDate(startDate)
-                .WithEditRouteStatusState(builder => builder
-                .WithStatus(status))
-            .Build();
-
-        var journeyInstance = await CreateJourneyInstanceAsync(
-            qualificationid,
-            editRouteState
-            );
-
-        var request = new HttpRequestMessage(HttpMethod.Post, $"/route/{qualificationid}/edit/end-date?{journeyInstance.GetUniqueIdQueryParameter()}")
-        {
-            Content = new FormUrlEncodedContentBuilder()
-            {
-                { "TrainingEndDate.Day", $"{endDate:%d}" },
-                { "TrainingEndDate.Month", $"{endDate:%M}" },
-                { "TrainingEndDate.Year", $"{endDate:yyyy}" },
-            }
-        };
-
-        // Act
-        var response = await HttpClient.SendAsync(request);
-
-        // Assert
-        Assert.Equal(StatusCodes.Status302Found, (int)response.StatusCode);
-        Assert.Equal($"/route/{qualificationid}/edit/holds-from?{journeyInstance.GetUniqueIdQueryParameter()}", response.Headers.Location?.OriginalString);
-        journeyInstance = await ReloadJourneyInstance(journeyInstance);
-        Assert.Equal(endDate, journeyInstance.State.EditStatusState!.TrainingEndDate);
-    }
-
-    [Fact]
-    public async Task Post_WhenNoEndDateIsEntered_ReturnsError()
-    {
-        // Arrange
-        var startDate = new DateOnly(2024, 01, 01);
-        var endDate = new DateOnly(2025, 01, 01);
-        var route = (await ReferenceDataCache.GetRouteToProfessionalStatusTypesAsync())
-            .Where(r => r.TrainingEndDateRequired == FieldRequirement.Mandatory)
-            .RandomOne();
-        var status = ProfessionalStatusStatusRegistry.All
-            .Where(s => s.TrainingEndDateRequired == FieldRequirement.Mandatory && s.HoldsFromRequired == FieldRequirement.NotApplicable)
-            .RandomOne()
-            .Value;
-        var person = await TestData.CreatePersonAsync(p => p
-            .WithRouteToProfessionalStatus(r => r
-                .WithRouteType(route.RouteToProfessionalStatusTypeId)
-                .WithStatus(status)));
-        var qualificationid = person.ProfessionalStatuses.First().QualificationId;
-        var editRouteState = new EditRouteStateBuilder()
-            .WithRouteToProfessionalStatusId(route.RouteToProfessionalStatusTypeId)
-            .WithStatus(status)
-            .WithTrainingStartDate(startDate)
-            .Build();
-
-        var journeyInstance = await CreateJourneyInstanceAsync(
-            qualificationid,
-            editRouteState
-            );
-
-        var request = new HttpRequestMessage(HttpMethod.Post, $"/route/{qualificationid}/edit/end-date?{journeyInstance.GetUniqueIdQueryParameter()}")
-        {
-            Content = new FormUrlEncodedContent(new Dictionary<string, string>())
-        };
-
-        // Act
-        var response = await HttpClient.SendAsync(request);
-
-        // Assert
-        await AssertEx.HtmlResponseHasErrorAsync(response, "TrainingEndDate", "Enter an end date");
-    }
-
-    [Fact]
-    public async Task Cancel_DeletesJourneyAndRedirectsToExpectedPage()
-    {
-        // Arrange
-        var startDate = new DateOnly(2024, 01, 01);
-        var endDate = new DateOnly(2025, 01, 01);
-        var route = (await ReferenceDataCache.GetRouteToProfessionalStatusTypesAsync())
-            .Where(r => r.TrainingEndDateRequired == FieldRequirement.Mandatory)
-            .RandomOne();
-        var status = ProfessionalStatusStatusRegistry.All
-            .Where(s => s.TrainingEndDateRequired == FieldRequirement.Mandatory)
-            .RandomOne()
-            .Value;
-        var person = await TestData.CreatePersonAsync(p => p
-            .WithRouteToProfessionalStatus(r => r
-                .WithRouteType(route.RouteToProfessionalStatusTypeId)
-                .WithStatus(RouteToProfessionalStatusStatus.Deferred)));
-        var qualificationid = person.ProfessionalStatuses.First().QualificationId;
-        var editRouteState = new EditRouteStateBuilder()
-            .WithRouteToProfessionalStatusId(route.RouteToProfessionalStatusTypeId)
-            .WithStatus(RouteToProfessionalStatusStatus.Deferred)
-            .WithTrainingStartDate(startDate)
-            .Build();
-
-        var journeyInstance = await CreateJourneyInstanceAsync(
-            qualificationid,
-            editRouteState
-            );
-
-        var request = new HttpRequestMessage(HttpMethod.Get, $"/route/{qualificationid}/edit/end-date?{journeyInstance.GetUniqueIdQueryParameter()}");
-
-        // Act
-        var response = await HttpClient.SendAsync(request);
-
-        // Assert
-        var doc = await AssertEx.HtmlResponseAsync(response);
-        var cancelButton = doc.GetElementByTestId("cancel-button") as IHtmlButtonElement;
-
-        // Act
-        var redirectRequest = new HttpRequestMessage(HttpMethod.Post, cancelButton!.FormAction);
-        var redirectResponse = await HttpClient.SendAsync(redirectRequest);
-
-        // Assert
-        Assert.Equal(StatusCodes.Status302Found, (int)redirectResponse.StatusCode);
-        var location = redirectResponse.Headers.Location?.OriginalString;
-        Assert.Equal($"/persons/{person.PersonId}/qualifications", location);
-        Assert.Null(await ReloadJourneyInstance(journeyInstance));
-    }
-
-    private Task<JourneyInstance<EditRouteState>> CreateJourneyInstanceAsync(Guid qualificationId, EditRouteState? state = null) =>
-        CreateJourneyInstance(
-           JourneyNames.EditRouteToProfessionalStatus,
-           state ?? new EditRouteState(),
-           new KeyValuePair<string, object>("qualificationId", qualificationId));
-}
